Move load_collection_name to new module.
