+++ src/antsibull_docs/lint_extra_docs.py
@@ -1,16 +1,13 @@
 # coding: utf-8
-# Author: Felix Fontein <tkuratom@redhat.com>
+# Author: Felix Fontein <felix@fontein.de>
 # License: GPLv3+
 # Copyright: Ansible Project, 2021
 """Lint extra collection documentation in docs/docsite/."""
 
-import json
 import os
 import os.path
 import re
 import typing as t
-
-from antsibull_core.yaml import load_yaml_file
 
 from .extra_docs import (
     find_extra_docs,
@@ -20,28 +17,12 @@ from .extra_docs import (
 )
 from .rstcheck import check_rst_content
 
+from .lint_helpers import (
+    load_collection_name,
+)
+
 
 _RST_LABEL_DEFINITION = re.compile(r'''^\.\. _([^:]+):''')
-
-
-def load_collection_name(path_to_collection: str) -> str:
-    '''Load collection name (namespace.name) from collection's galaxy.yml.'''
-    manifest_json_path = os.path.join(path_to_collection, 'MANIFEST.json')
-    if os.path.isfile(manifest_json_path):
-        with open(manifest_json_path, 'rb') as f:
-            manifest_json = json.load(f)
-        # pylint:disable-next=consider-using-f-string
-        collection_name = '{namespace}.{name}'.format(**manifest_json['collection_info'])
-        return collection_name
-
-    galaxy_yml_path = os.path.join(path_to_collection, 'galaxy.yml')
-    if os.path.isfile(galaxy_yml_path):
-        galaxy_yml = load_yaml_file(galaxy_yml_path)
-        # pylint:disable-next=consider-using-f-string
-        collection_name = '{namespace}.{name}'.format(**galaxy_yml)
-        return collection_name
-
-    raise Exception(f'Cannot find files {manifest_json_path} and {galaxy_yml_path}')
 
 
 # pylint:disable-next=unused-argument
